Added simulation scripts, single-core submission script.
The assembly conditions are in the name.  This specific instance used 540 viruses at cV = 370 nM, amidst 100x dendrimer.

The assembly.pbs script can be optimized for BR3, using greater parallelism given we run fewer jobs for this project than the previous.  It is not going to represent actual time to completion well as-is, but should run.